% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEGsPathwayVisualization.R
\name{DEGsPathwayVisualization}
\alias{DEGsPathwayVisualization}
\title{The function uses the DEGs to show graphically the expression of the samples in the pathways in which those genes appear.}
\usage{
DEGsPathwayVisualization(DEGsMatrix, DEGsAnnotation, expressionMatrix,
  expressionAnnotation, labels)
}
\arguments{
\item{DEGsMatrix}{A matrix that contains the expression of the DEGs for each samples. This matrix can be achieved by calling the function \code{\link{limmaDEGsExtraction}}. If the samples are more than 24, only the first 24 will be used to colour the pathways.}

\item{DEGsAnnotation}{A matrix that contains the gene names and the entrez IDs for the genes available in the DEGs matrix. This annotation can be obtained from the function \code{\link{getAnnotationFromEnsembl}}.}

\item{expressionMatrix}{A matrix that contains the expression of the all the genes available for each samples. If the samples are more than 24, only the first 24 will be used to colour the pathways.}

\item{expressionAnnotation}{A matrix that contains the gene names and the entrez IDs for all the genes available. This annotation can be obtained from the function \code{\link{getAnnotationFromEnsembl}}.}

\item{labels}{A vector that contains the labels of the samples for both the DEGsMatrix and the expressionMatrix.}
}
\value{
Nothing to return.
}
\description{
The function uses the DEGs to show graphically the expression of the samples in the pathways in which those genes appear. For that, the function makes use of a DEGsMatrix with the expression of the DEGs and the annotation of those DEGs in which appear the pathway or pathways of each DEGs. Internally, the function uses \code{\link{pathview}} to retrieve and colours the pathways, but a maximum number of 24 samples can be used. Furthermore, the function needs the expression matrix with all the genes in order to use them to colour the rest of the elements in the pathways.
}
\examples{
downloadPublicSeries(c("GSE74251","GSE81593"))

GSE74251 <- read.csv("ReferenceFiles/GSE74251.csv")
GSE81593 <- read.csv("ReferenceFiles/GSE81593.csv")

GSE74251 <- GSE74251[1:5,]
GSE81593 <- GSE81593[8:12,]

dir <- system.file("extdata", package="KnowSeq")

Run <- GSE74251$Run
Path <- paste(dir,"/countFiles/",GSE74251$Run,sep = "")
Class <- rep("Tumor", length(GSE74251$Run))
GSE74251CountsInfo <-  data.frame(Run = Run, Path = Path, Class = Class)

Run <- GSE81593$Run
Path <- paste(dir,"/countFiles/",GSE81593$Run,sep = "")
Class <- rep("Control", length(GSE81593$Run))
GSE81593CountsInfo <-  data.frame(Run = Run, Path = Path, Class = Class)

mergedCountsInfo <- rbind(GSE74251CountsInfo, GSE81593CountsInfo)

write.csv(mergedCountsInfo, file = "ReferenceFiles/mergedCountsInfo.csv")

countsInformation <- countsToMatrix("ReferenceFiles/mergedCountsInfo.csv")

countsMatrix <- countsInformation$countsMatrix
labels <- countsInformation$labels

myAnnotation <- getAnnotationFromEnsembl(rownames(countsMatrix),referenceGenome=37)

expressionMatrix <- calculateGeneExpressionValues(countsMatrix,myAnnotation, genesNames = TRUE)

DEGsInformation <- limmaDEGsExtraction(expressionMatrix, labels, lfc = 2.0,
pvalue = 0.01, number = Inf)

topTable <- DEGsInformation$Table

DEGsMatrix <- DEGsInformation$DEGsMatrix

myDEGsAnnotation <- getAnnotationFromEnsembl(rownames(DEGsMatrix)[1:3],
referenceGenome=38,attributes = c("external_gene_name","entrezgene"),
filters = "external_gene_name")

allMyAnnotation <-  getAnnotationFromEnsembl(rownames(expressionMatrix),
                                             referenceGenome=38,attributes = c("external_gene_name","entrezgene"),
                                             filters = "external_gene_name")

DEGsPathwayVisualization(DEGsMatrix[1:3,], myDEGsAnnotation, expressionMatrix, allMyAnnotation, labels)
}
