% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEGsToDiseases.R
\name{DEGsToDiseases}
\alias{DEGsToDiseases}
\title{DEGsToDiseases obtains the information about what diseases are related to the DEGs indicated by parameter.}
\usage{
DEGsToDiseases(geneList, minCitation = 5)
}
\arguments{
\item{geneList}{A list that contains the gene symbols or gene names of the DEGs.}

\item{minCitation}{Minimum number of citations of each genes in a disease to consider the genes related with the disease.}
}
\value{
A list which contains the information about the diseases associated to each genes and to a sets of genes.
}
\description{
The function obtains the information about what diseases are related to the DEGs indicated by parameter. For that, the function makes use of the web platform gene2Diseases.
}
\examples{
diseases <- DEGsToDiseases(c("KRT19","BRCA1"))
}
