% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEGsToPathways.R
\name{DEGsToPathways}
\alias{DEGsToPathways}
\title{The function uses the DEGs to retrieves the different pathways in which those DEGs involve any interaction.}
\usage{
DEGsToPathways(geneList)
}
\arguments{
\item{geneList}{A list which contains the DEGs that will be used to retrieve the related pathways to them.}
}
\value{
A list with the pathways that contain relation to the DEGs within the geneList parameter.
}
\description{
The function uses the DEGs to retrieves the different pathways in which those DEGs involve any interaction.
}
\examples{
DEGsToPathways(c("BRCA1","MLANA"))

}
