% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadPublicSeries.R
\name{downloadPublicSeries}
\alias{downloadPublicSeries}
\title{Download automatically samples from NCBI/GEO and ArrayExpress public databases.}
\usage{
downloadPublicSeries(samplesVector)
}
\arguments{
\item{samplesVector}{A vector which contains the different IDs of the wanted series. These IDs are the IDs of the series from NCBI/GEO or ArrayExpress.}
}
\value{
Nothing to return.
}
\description{
Download automatically samples from series of either microarray and RNA-seq. Furthermore, both NCBI/GEO and ArrayExpress public databases are supported. In the case of Microarray, the raw file are downloaded, if they are available, but for RNA-seq a csv is created with the necessary information to download the samples with the function \code{\link{rawAlignment}}.
}
\examples{
downloadPublicSeries(c("GSE74251","GSE81593"))
}
