% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileMove.R
\name{fileMove}
\alias{fileMove}
\title{This function is used to move files to other locations.}
\usage{
fileMove(from, to)
}
\arguments{
\item{from}{The current path to the file.}

\item{to}{The path to the new location of the file.}
}
\value{
nothing to return
}
\description{
This function is used to move files to other locations.
}
\examples{
fileMove("~/source", "~/newLocation")
}
