% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdcClientDownload.R
\name{gdcClientDownload}
\alias{gdcClientDownload}
\title{This function downloads a list of controlled files from GDC Portal with the user token and the manifest with the information about the desired controlled files.}
\usage{
gdcClientDownload(tokenPath, manifestPath, data)
}
\arguments{
\item{tokenPath}{Path to the GDC token}

\item{manifestPath}{Path to the samples manifest}

\item{data}{The matrix or data.frame with the information from the Samples Sheet downloaded from GDC Portal.}
}
\value{
Nothing to return.
}
\description{
This function downloads a list of controlled files from GDC Portal with the user token and the manifest with the information about the desired controlled files.
}
