% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAnnotationFromEnsembl.R
\name{getAnnotationFromEnsembl}
\alias{getAnnotationFromEnsembl}
\title{getAnnotationFromEnsembl returns the required information about a list of genes from Ensembl biomart.}
\usage{
getAnnotationFromEnsembl(
  values,
  attributes = c("ensembl_gene_id", "external_gene_name", "percentage_gene_gc_content",
    "gene_biotype"),
  filters = "ensembl_gene_id",
  referenceGenome = 38,
  notHSapiens = FALSE,
  notHumandataset = ""
)
}
\arguments{
\item{values}{A list of genes that contains the names or IDs.}

\item{attributes}{A vector which contains the different information attributes that the Ensembl biomart admit.}

\item{filters}{The attributes used as filter to return the rest of the attributes.}

\item{referenceGenome}{The human reference genome used to return the annotation. The possibilities are two: 37 and 38}

\item{notHSapiens}{A boolean value that indicates if the user wants the human annotation or another annotation available in BiomaRt. The possible not human dataset can be consulted by calling the following function: biomaRt::listDatasets(useMart("ensembl")).}

\item{notHumandataset}{A dataset identification from biomaRt::listDatasets(useMart("ensembl")).}
}
\value{
A matrix that contains all the information asked to the attributes parameter.
}
\description{
The function returns the required information about a list of genes from Ensembl biomart. This list of genes can be Ensembl ID, gene names or either of the possible values admited by Ensembl biomart. Furthermore, the reference genome can be chosen depending on the necessity of the user.
}
\examples{
myAnnotation <- getAnnotationFromEnsembl(c("ENSG00000210049","ENSG00000211459","ENSG00000210077"),referenceGenome=37)
}
