% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knowseqReport.R
\name{knowseqReport}
\alias{knowseqReport}
\title{knowseqReport creates a report for a given set of genes which their label.}
\usage{
knowseqReport(
  data,
  labels,
  outdir = "knowSeq-report",
  baseline = "expression",
  featureSelectionMode = "mrmr",
  disease = "",
  maxGenes = 12,
  clasifAlgs = c("knn", "rf", "svm"),
  metrics = c("accuracy", "specificity", "sensitivity")
)
}
\arguments{
\item{data}{A matrix that contains the gene expression or counts values.}

\item{labels}{A vector or factor that contains the labels for each of the samples in the data object.}

\item{outdir}{The output directory to store the report.}

\item{baseline}{A string that indicates the start point. This will be 'expression' if data contains genes expression values or 'counts' if data contains genes counts values.}

\item{featureSelectionMode}{String that indicates which feature selection algorithm is going to be used. Possible values are: mrmr, rf or da.}

\item{disease}{String that indicates from which disease wants the user to know if selected genes are related to. Found evidences will be shown. Default empty, this means that all related diseases, and found evidences, will be shown.}

\item{maxGenes}{Integer that indicates the maximun number of genes which information will be shown and that will be used to train models.}

\item{clasifAlgs}{A vector with including algorithms names that will be used in training cv.}
}
\value{
Nothing to return.
}
\description{
knowseqReport creates a report for a given set of genes which their label. This provide an html file with all the information that can be obtained for a certain set of genes (as GO, pathway visualization, associated diseases) and their labels (machine learning process).
}
\examples{
dir <- system.file("extdata", package="KnowSeq")
load(paste(dir,"/expressionExample.RData",sep = ""))
knowseqReport(expressionMatrix,labels,'knowSeq-report',clasifAlgs=c('rf'),disease='cancer',maxGenes = 9)
}
