% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotConfMatrix.R
\name{plotConfMatrix}
\alias{plotConfMatrix}
\title{plotConfMatrix plots a confusion matrix with some statistics.}
\usage{
plotConfMatrix(data)
}
\arguments{
\item{data}{A table which contains a confusion matrix.}
}
\value{
Nothing to return.
}
\description{
The function plots a confusion matrix with some statistics. The function is used internally by \code{\link{dataPlot}} but it can be used separatelly.
}
\examples{
data <- table(as.factor(c(1,2,4,2,4,5)),as.factor(c(1,2,5,4,5,2)))
plotConfMatrix(data)
}
