% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAseqQA.R
\name{RNAseqQA}
\alias{RNAseqQA}
\title{RNAseqQA performs the quality analysis of an expression matrix.}
\usage{
RNAseqQA(expressionMatrix, outdir = "RNAseqQA")
}
\arguments{
\item{expressionMatrix}{A matrix that contains the gene expression values.}

\item{outdir}{The output directory to store the report of arrayQualityMetrics}
}
\value{
Nothing to return.
}
\description{
RNAseqQA performs the quality analysis of an expression matrix. This function adapts the RNA-seq data in order to allows using arrayQualityMetrics expression analysis.
}
\examples{
downloadPublicSeries(c("GSE74251","GSE81593"))

GSE74251 <- read.csv("ReferenceFiles/GSE74251.csv")
GSE81593 <- read.csv("ReferenceFiles/GSE81593.csv")

GSE74251 <- GSE74251[1:5,]
GSE81593 <- GSE81593[8:12,]

dir <- system.file("extdata", package="KnowSeq")

Run <- GSE74251$Run
Path <- paste(dir,"/countFiles/",GSE74251$Run,sep = "")
Class <- rep("Tumor", length(GSE74251$Run))
GSE74251CountsInfo <-  data.frame(Run = Run, Path = Path, Class = Class)

Run <- GSE81593$Run
Path <- paste(dir,"/countFiles/",GSE81593$Run,sep = "")
Class <- rep("Control", length(GSE81593$Run))
GSE81593CountsInfo <-  data.frame(Run = Run, Path = Path, Class = Class)

mergedCountsInfo <- rbind(GSE74251CountsInfo, GSE81593CountsInfo)

write.csv(mergedCountsInfo, file = "ReferenceFiles/mergedCountsInfo.csv")

countsInformation <- countsToMatrix("ReferenceFiles/mergedCountsInfo.csv")

countsMatrix <- countsInformation$countsMatrix
labels <- countsInformation$labels

myAnnotation <- getAnnotationFromEnsembl(rownames(countsMatrix),referenceGenome=37)

expressionMatrix <- calculateGeneExpressionValues(countsMatrix,myAnnotation, genesNames = TRUE)

RNAseqQA(expressionMatrix)
}
