% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sraToFastq.R
\name{sraToFastq}
\alias{sraToFastq}
\title{sraToFastq downloads and converts the sra files to fastq files. The function admits both gz and sra formats.}
\usage{
sraToFastq(identifier)
}
\arguments{
\item{identifier}{A vector that contains a list with the URLs requested.}
}
\value{
Nothing.
}
\description{
This function downloads and converts the sra files to fastq files by using the URLs indicated through the identifier argument. The function admits both gz and sra formats. This function is used internally by \code{\link{rawAlignment}} but it can be used separatelly.
}
\examples{
# This function needs the download of the pre-compiled tools supplied by KnowSeq.

\dontrun{sraToFastq("SRA1")}
}
