% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchEffectRemoval.R
\name{batchEffectRemoval}
\alias{batchEffectRemoval}
\title{Corrects the batch effect of the data by using the selected method.}
\usage{
batchEffectRemoval(expressionMatrix, labels, method = "combat",
  clusters = 2)
}
\arguments{
\item{expressionMatrix}{The original expression matrix to treat the batch effect.}

\item{labels}{A vector that contains the labels of the samples in expressionMatrix.}

\item{method}{The method that will be used to remove the batch effect. The possibilities are "combat" or "sva". Next release will add RUV.}

\item{clusters}{The number of clusters intrinsic to the expression matrix data which could means different batches. The optimal number of clusters in the expression matrix can be calculated by calling the function \code{\link{dataPlot}}, with the parameter mode equal to "optimalClusters". This parameter is only required when the user selects the combat method.}
}
\value{
A matrix with the batch effect corrected for combat or a model for \code{\link{limmaDEGsExtraction}} function in the case of sva.
}
\description{
This function corrects the batch effect of the expression matrix indicated by parameter. There are two method to choose such as ComBat or SVA.
}
\examples{
downloadPublicSeries(c("GSE74251","GSE81593"))

GSE74251 <- read.csv("ReferenceFiles/GSE74251.csv")
GSE81593 <- read.csv("ReferenceFiles/GSE81593.csv")

GSE74251 <- GSE74251[1:5,]
GSE81593 <- GSE81593[8:12,]

dir <- system.file("extdata", package="KnowSeq")

Run <- GSE74251$Run
Path <- paste(dir,"/countFiles/",GSE74251$Run,sep = "")
Class <- rep("Tumor", length(GSE74251$Run))
GSE74251CountsInfo <-  data.frame(Run = Run, Path = Path, Class = Class)

Run <- GSE81593$Run
Path <- paste(dir,"/countFiles/",GSE81593$Run,sep = "")
Class <- rep("Control", length(GSE81593$Run))
GSE81593CountsInfo <-  data.frame(Run = Run, Path = Path, Class = Class)

mergedCountsInfo <- rbind(GSE74251CountsInfo, GSE81593CountsInfo)

write.csv(mergedCountsInfo, file = "ReferenceFiles/mergedCountsInfo.csv")

countsInformation <- countsToMatrix("ReferenceFiles/mergedCountsInfo.csv")

countsMatrix <- countsInformation$countsMatrix
labels <- countsInformation$labels

myAnnotation <- getAnnotationFromEnsembl(rownames(countsMatrix),referenceGenome=37)

expressionMatrix <- calculateGeneExpressionValues(countsMatrix,myAnnotation, genesNames = TRUE)

expressionMatrixNoBatch <- batchEffectRemoval(expressionMatrix, labels, clusters = 4)
svaMod <- batchEffectRemoval(expressionMatrix, labels, method = "sva")
}
