% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdcClientDownload.R
\name{gdcClientDownload}
\alias{gdcClientDownload}
\title{This function downloads a list of controlled files from GDC Portal with the user token and the manifest with the information about the desired controlled files.}
\usage{
gdcClientDownload(manifestPath, controlled = FALSE, tokenPath = "")
}
\arguments{
\item{manifestPath}{Path to the samples manifest}

\item{controlled}{Parameter that indicates if data to download are controlled or not}

\item{tokenPath}{Path to the GDC token if data are controlled}
}
\value{
Nothing to return.
}
\description{
This function downloads a list of controlled files from GDC Portal with the user token and the manifest with the information about the desired controlled files.
}
\examples{
# This function needs the download of the pre-compiled tools supplied by KnowSeq.
\dontrun{gdcClientDownload("PathToTheToken", "PathToTheFileWithDownloadInfo", dataMatrix)}
}
