% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneOntologyEnrichment.R
\name{geneOntologyEnrichment}
\alias{geneOntologyEnrichment}
\title{geneOntologyEnrichment obtains the information about what Gene Ontology terms are related to the DEGs.}
\usage{
geneOntologyEnrichment(
  geneMatrix,
  labels,
  identificator = "SYMBOL",
  mapping = "org.Hs.eg.db",
  nGOs = 10,
  pvalCutOff = 0.01
)
}
\arguments{
\item{geneMatrix}{A matrix that contains the expression of the DEGs for each samples.}

\item{labels}{A vector that contains the labels of the samples of the DEGsMatrix.}

\item{identificator}{The identification methods for the genes. By default the identificator is the gene symbol or gene name.}

\item{mapping}{The annotation database to map the gene and GOs. By default is prepared for homo sapiens but can be changed for other species.}

\item{nGOs}{Maximun number of GOs to return of the total amount of top GOs. By default is equal to 10.}

\item{pvalCutOff}{The maximum p-value to considers that a genes is related with a GO term.}
}
\value{
A list that contains a matrix for each of the possible ontologies and a matrix with the GOs for the three ontologies together.
}
\description{
The function obtains the information about GO terms from the three differents ontologies that are related to the DEGs. The function also returns the description about each GO and a list of genes that are inside of each GO.
}
\examples{
dir <- system.file("extdata", package="KnowSeq")
load(paste(dir,"/expressionExample.RData",sep = ""))

labelsGo <- gsub("Control",0,labels)
labelsGo <- gsub("Tumor",1,labelsGo)

GOsList <- geneOntologyEnrichment(DEGsMatrix,labelsGo,nGOs = 20,pvalCutOff = 0.001)
}
