% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_test.R
\name{knn_test}
\alias{knn_test}
\title{knn_test allows assessing the final DEGs through a machine learning step by using k-NN with a test dataset.}
\usage{
knn_test(train, labelsTrain, test, labelsTest, vars_selected)
}
\arguments{
\item{train}{The train parameter is an expression matrix or data.frame that contains the training dataset with the genes in the columns and the samples in the rows.}

\item{labelsTrain}{A vector or factor that contains the training labels for each of the samples in the train object.}

\item{test}{The test parameter is an expression matrix or data.frame that contains the test dataset with the genes in the columns and the samples in the rows.}

\item{labelsTest}{A vector or factor that contains the test labels for each of the samples in the test object.}

\item{vars_selected}{The genes selected to classify by using them. It can be the final DEGs extracted with the function \code{\link{limmaDEGsExtraction}} or a custom vector of genes. Furthermore, the ranking achieved by \code{\link{featureSelection}} function can be used as input of this parameter.}
}
\value{
A list that contains four objects. The confusion matrix, the accuracy, the sensitibity and the specificity for each genes.
}
\description{
knn_test allows assessing the final DEGs through a machine learning step by using k-NN with a test dataset. An optimization of the k neighbours is done at the start of the process.
}
\examples{
knn_test(DEGsMatrix,labelsTrain,test,labelsTest,rownames(DEGsMatrix)[1:10])
knn_test(DEGsMatrix,labelsTrain,test,labelsTest,rownames(DEGsMatrix))
}
