% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAseqQA.R
\name{RNAseqQA}
\alias{RNAseqQA}
\title{RNAseqQA performs the quality analysis of an expression matrix.}
\usage{
RNAseqQA(expressionMatrix, outdir = "RNAseqQA")
}
\arguments{
\item{expressionMatrix}{A matrix that contains the gene expression values.}

\item{outdir}{The output directory to store the report of arrayQualityMetrics}
}
\value{
Nothing to return.
}
\description{
RNAseqQA performs the quality analysis of an expression matrix. This function adapts the RNA-seq data in order to allows using arrayQualityMetrics expression analysis.
}
\examples{
dir <- system.file("extdata", package="KnowSeq")
load(paste(dir,"/expressionExample.RData",sep = ""))

RNAseqQA(expressionMatrix)
}
